Change repo to only build with travis on nightly
While GCP is what hosts the website, it's still good to know when things
are breaking. This builds the code before hand on non master branches.
If it works we can say everything went fine then actually push to
master, whereby GCP will build the docker image and deploy. Since we use
linux we only need to test for it and can ignore appveyor completely.
This also fixes an issue where we needed rocket on nightly and not
stable.