Implement the base site
This sets up a site one can navigate too. It contains all of the mailing
lists, some basic css, a contact page and the index page. It's all based
off my site mgattozzi.com and has much room for improvement but this is
a good start and allows people to view this all online.